% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n.R
\name{get_n}
\alias{get_n}
\title{Identify if it has a study}
\usage{
get_n(metadatdat, n_var = NULL)
}
\arguments{
\item{metadatdat}{A dataset from \link{metadat}.}

\item{n_var}{Specify which column which has studies.}
}
\value{
Vector of sample sizes in a \link{metadat} dataset.
}
\description{
Identify if it has a study
}
