% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_of_dat.R
\docType{data}
\name{meta_of_dat}
\alias{meta_of_dat}
\title{Meta data of the metadat:: datasets}
\format{
An object of class \code{data.frame} with 55 rows and 5 columns.
}
\usage{
meta_of_dat
}
\description{
Imported metadata. See \emph{data-raw/meta_of_dat.R}.
}
\keyword{datasets}
