% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_k.R
\name{get_k}
\alias{get_k}
\title{Return number of studies in a meta-analysis}
\usage{
get_k(metadatdat)
}
\arguments{
\item{metadatdat}{A dataset from \link{metadat}.}
}
\value{
Number of studies, k. Or "no study id" if we couldn't find one.
}
\description{
Return number of studies in a meta-analysis
}
